requirements.txt
aiohttp==3.8.0
aiosignal==1.2.0
appdirs==1.4.4
asgiref==3.3.4
astroid==2.5.1
async-timeout==4.0.0
attrs==21.2.0
Automat==20.2.0
autopep8==1.5.7
beautifulsoup4==4.9.3
cbor2==5.4.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.7
colorama==0.4.4
configparser==5.0.2
constantly==15.1.0
crayons==0.4.0
cryptography==3.4.7
cssselect==1.1.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==3.10.0
django-heroku==0.3.1
django-phone-field==1.8.1
django-phonenumber-field==5.2.0
django-recaptcha==2.0.6
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
docopt==0.6.2
filelock==3.0.12
frozenlist==1.2.0
gunicorn==20.1.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
isort==5.8.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
lazy-object-proxy==1.6.0
lxml==4.6.3
mccabe==0.6.1
multidict==5.2.0
O365==2.0.15
oauthlib==3.1.1
parsel==1.6.0
pipenv==2020.11.15
pipreqs==0.4.10
priority==1.3.0
Protego==0.1.16
psycopg2==2.8.6
psycopg2-binary==2.9.1
pubnub==5.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pycryptodomex==3.11.0
PyDispatcher==2.0.5
PyJWT==2.3.0
pylint==2.7.2
pyOpenSSL==20.0.1
python-dateutil==2.8.2
pytz==2021.1
queuelib==1.6.1
requests==2.25.1
requests-oauthlib==1.3.0
Scrapy==2.5.0
selenium==3.141.0
service-identity==21.1.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
stringcase==1.2.0
stripe==2.61.0
toml==0.10.2
twilio==6.58.0
Twisted==21.2.0
typing-extensions==3.10.0.2
tzlocal==2.1
urllib3==1.26.4
virtualenv==20.4.6
virtualenv-clone==0.5.4
w3lib==1.22.0
webdriver-manager==3.3.0
whitenoise==5.2.0
wrapt==1.12.1
yarg==0.1.9
yarl==1.7.0
zope.interface==5.4.0
